+++ max.h
@@ -11,8 +11,8 @@
 
 class Max: public Operator {
 public:
-    void operate(Result& result, const std::vector<uint16_t>& data,
-                 const size_t& from, const size_t& to) const override;
+    void operate(Result & result, const std::vector<uint16_t> & data,
+                 const uint32_t & from, const uint32_t & to) const override;
 };
